pyproject.toml
[tool.poetry]
name = "landslide4sense"
version = "0.0.1"
description = "Detecting landslide with AI"
authors = ["sebastiangarciaacosta <segaracos@outlook.com>"]

[tool.poetry.dependencies]
pip = "22.0.4"
setuptools = "58.0.4"
python = ">=3.7,<3.11"
numpy = "^1.21.5"
wandb = "^0.12.11"
torch = "1.10.0"
h5py = "3.1.0"
albumentations= "1.1.0"
ipykernel = "^6.12.1"
jupyter = "^1.0.0"
hydra-core = "^1.1.1"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]
pip-tools = "^6.6.0"
black = "^22.3.0"
google-api-python-client = "^2.43.0"
google-cloud-storage = "^2.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
